style: Edit the resposive style of the main and restaurant details pages
Add an extra break point to the main page. Change the restaurants list style. Fix the fonts to be responsive.

In the restaurant details page change the review, the restaurant's image and the map style. Fix the fonts. Group the writer, the date and the rating of the review in one container.